requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
cuid==0.4
Django==6.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
